Move transparency buffer to the geometry folder
