+++ trview.app/Geometry/TransparencyBuffer.cpp
