+++ trview.app/Geometry/TransparencyBuffer.h
