+++ trview.app/trview.app.vcxproj
@@ -35,6 +35,7 @@
     <ClCompile Include="Flipmaps.cpp" />
     <ClCompile Include="Geometry\IRenderable.cpp" />
     <ClCompile Include="Geometry\Mesh.cpp" />
+    <ClCompile Include="Geometry\TransparencyBuffer.cpp" />
     <ClCompile Include="Geometry\TransparentTriangle.cpp" />
     <ClCompile Include="GoTo.cpp" />
     <ClCompile Include="ILevelTextureStorage.cpp" />
@@ -58,7 +59,6 @@
     <ClCompile Include="SelectionRenderer.cpp" />
     <ClCompile Include="SettingsWindow.cpp" />
     <ClCompile Include="Toolbar.cpp" />
-    <ClCompile Include="TransparencyBuffer.cpp" />
     <ClCompile Include="Trigger.cpp" />
     <ClCompile Include="TriggersWindow.cpp" />
     <ClCompile Include="TriggersWindowManager.cpp" />
@@ -87,6 +87,7 @@
     <ClInclude Include="Geometry\IRenderable.h" />
     <ClInclude Include="Geometry\Mesh.h" />
     <ClInclude Include="Geometry\MeshVertex.h" />
+    <ClInclude Include="Geometry\TransparencyBuffer.h" />
     <ClInclude Include="Geometry\TransparentTriangle.h" />
     <ClInclude Include="Geometry\Triangle.h" />
     <ClInclude Include="GoTo.h" />
@@ -113,7 +114,6 @@
     <ClInclude Include="SelectionRenderer.h" />
     <ClInclude Include="SettingsWindow.h" />
     <ClInclude Include="Toolbar.h" />
-    <ClInclude Include="TransparencyBuffer.h" />
     <ClInclude Include="Trigger.h" />
     <ClInclude Include="TriggersWindow.h" />
     <ClInclude Include="TriggersWindowManager.h" />
