+++ trview/Level.cpp
@@ -10,7 +10,7 @@
 #include "LevelTextureStorage.h"
 #include "MeshStorage.h"
 #include <trview.app/Camera/ICamera.h>
-#include <trview.app/TransparencyBuffer.h>
+#include <trview.app/Geometry/TransparencyBuffer.h>
 #include "ResourceHelper.h"
 #include "resource.h"
 #include <trview.app/SelectionRenderer.h>
