+++ trview/Room.h
@@ -16,7 +16,7 @@
 #include <trlevel/ILevel.h>
 
 #include "StaticMesh.h"
-#include <trview.app/TransparencyBuffer.h>
+#include <trview.app/Geometry/TransparencyBuffer.h>
 #include <trview.app/Geometry/Mesh.h>
 #include <trview.app/Sector.h>
 #include <trview.app/PickResult.h>
