+++ trview/StaticMesh.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "StaticMesh.h"
 #include <trview.app/Geometry/Mesh.h>
-#include <trview.app/TransparencyBuffer.h>
+#include <trview.app/Geometry/TransparencyBuffer.h>
 
 namespace trview
 {
